requirements.txt
pip>=23.1
setuptools>=65.0.0
wheel

flask
streamlit
pymupdf
unstructured[local-inference]
python-magic
filetype

beautifulsoup4
chromadb
sentence-transformers
langchain
langchain-text-splitters
faiss-cpu
python-dotenv
selenium



# 🆕 Add Gemini support
google-generativeai
